requirements.txt
fastapi
uvicorn
PyYAML
pytest
requests
